Clear Temp Files: add --test flag

+++ bin/clear_old_temp_files.sh
@@ -7,9 +7,14 @@
 #
 # Arguments:
 #   first = the number of days as the max age of the temp files (7 by default)
+#   --test = look for files to delete, but do not actually delete them
 #
 
-MAX_AGE_DAYS=${1:-7}
+DEFAULT_MAX_AGE_DAYS=7
+MAX_AGE_DAYS=${1:-$DEFAULT_MAX_AGE_DAYS}
+if [[ $MAX_AGE_DAYS == '--test' ]]; then
+  MAX_AGE_DAYS=$DEFAULT_MAX_AGE_DAYS
+fi
 SGN_LOCAL_CONF="/home/production/cxgn/sgn/sgn_local.conf"
 SGN_CONF="/home/production/cxgn/sgn/sgn.conf"
 TMP_DIR=""
@@ -42,15 +47,23 @@ fi
 
 # Find old files in tmp directory
 echo "--> finding temp files older than $MAX_AGE_DAYS days in $TMP_DIR..."
-files=$(find "$TMP_DIR" -mtime +$MAX_AGE_DAYS -type f -ls)
+if [[ $MAX_AGE_DAYS != 0 ]]; then
+  files=$(find "$TMP_DIR" -mtime +$MAX_AGE_DAYS -type f -ls)
+else
+  files=$(find "$TMP_DIR" -type f -ls)
+fi
 
 echo "$files" | while read line; do
   filename=$(echo "$line" | awk '{print $NF}')
   if [ ! -z "$filename" ] && [ -e "$filename" ]; then
     base=$(basename $filename)
     if [[ "$base" != git_* ]]; then
-      echo "--> removing $filename"
-      rm "$filename"
+      if [[ "$1" != "--test" ]] && [[ "$2" != "--test" ]]; then
+        echo "--> removing $filename"
+        rm "$filename"
+      else
+        echo "--> would remove $filename"
+      fi
     fi
   fi
 done
